tests: force stdout flush after completing unit tests
Use a custom test suite instance when building unit tests to run to help
force the standard output stream to flush on completion. This helps
ensure that any buffered content on the standard output stream is
rendered before a unit test result is dumped to an invoker’s standard
error stream.

Signed-off-by: James Knight <james.d.knight@live.com>