tests: configure working directory to test environment's directory
When building a releng-tool test environment (`prepare_testenv` or
`run_testenv`), have the working context automatically change the
working directory to the interim directory created for the test. This is
be more flexible for unit tests who want to invoke I/O-specific calls in
the root/temp. directory for a test without having to manually move into
the directory itself.

Signed-off-by: James Knight <james.d.knight@live.com>