tests: provide a git-site tools tests
Provide a series of unit tests to help validate support for the use of
the Git tool in releng-tool. This commit adds a `TestSiteToolBase` base
class to help support site-specific tool verification. The first attempt
to help improving tool validation is done with the verification of a
series of Git capabilities -- specifically, the ability to fetch and
extract repositories of different revisions with various configurations.

Signed-off-by: James Knight <james.d.knight@live.com>