fetch-url: cleanup progress line on completion
A URL fetch will output the progress of a download on the same line
until completion. After the download, a "completed download" line is
printed to the standard output stream. Unfortunately, the download
completion line is printed on-top of the previous download progress,
which may result in a non-clean line on the output stream. To prevent
this, force the ending of the line by injecting a newline if any bytes
have been read.

Signed-off-by: James Knight <james.d.knight@live.com>