engine: allow pre-config package-specific extraction/patching
A user has the ability to fetch a single package in a project by using a
`<pkg>-fetch` command, without having a need to download other packages
(even if there is a dependency). The `extract` and `patch` actions are
not required to have dependencies processed as well; thus, adjusting
the package-variants of these commands to only invoke stages on the
provided package.

Signed-off-by: James Knight <james.d.knight@live.com>