io: introduce (releng_)execute_rv function
Adding an alternative execute function (`execute_rv`) to provide an easy
way to invoke a command for its return code/output. This allows a call
to react to specific return codes (if needed), as well as avoids the
need of setting up a capture buffer.

Also, exposing this function via `releng_execute_rv`, to provide this
flexibility for releng-tool scripts.

Signed-off-by: James Knight <james.d.knight@live.com>