io: convert none execute arguments to empty strings
Convert any execute-provided `None` argument to an empty string. While
execute arguments are not intended to accept `None` entries, a script
environment may accidentally pass in a `None` value which could result
in some unexpected outputs (specifically, the execute operations may
throw an exception without a clear understanding to the user on why).

Signed-off-by: James Knight <james.d.knight@live.com>