fetch-git: force clean invalid cache when ignoring cache checks
If a user requests a force-fetch on a package, some cache checks are
skipped to allow re-fetching information from the configured origin. In
the case where a cache directory is broken enough that Git operations
fail on it, a forced-fetch operation will always fail. In the event that
a broken Git cache is detected, wipe the cache for a clean re-fetch.

Signed-off-by: James Knight <james.d.knight@live.com>