util-env: allow env helper to support quick-fetching
Adjusting the `releng_env` script helper to support fetching an
environment variable's value as well, when no explicit value has been
provided. This should provide flexibility for users trying to get and
set environment variables without needing to import the `os` module
directly.

This commit also properly support the deletion of environment variables
when a `None` value is passed in.

Signed-off-by: James Knight <james.d.knight@live.com>