fetch-git: support lighter fetch for git
Adding a faster way to checkout a Git target by attempting to only
fetch a specific head/reference if they are detected as available on the
configured remote. This avoids the need to fetch all head/tag references
from the repository, which may not be needed for most users. In the
event that a specific head/reference cannot be found, it will fallback
to the original fetch heads/references attempt, followed by an unshallow
fetch attempt (if shallow fetches have not been disabled).

Signed-off-by: James Knight <james.d.knight@live.com>