fetch: support sharing dvcs caches
Provides a way to share DVCS caches between various packages.
Originally, a cache folder for a DVCS source would be held in a
`cache/<pkg-name>` location. While functional, the fetching time for a
project will increase when:

1) More than one package definition uses the same site.
2) (Git) Each individual package's submodules cache is re-fetched for
   each package.

To help minimize the time and space needed for these packages, this
commit provides a way for packages to auto-detect when caches are
shared and use the same cache directory. This is done with a combination
of hashing the site target to find common origins, as well as tracking
cache locations/keys for packages for clean pre-cached builds. Packages
will default to using the `cache/<pkg-name>` pattern; submodules will be
stored inside `cache/hash(<pkg-site>)`; and when a match is detected,
keep track of an alternative container to be used.

Signed-off-by: James Knight <james.d.knight@live.com>