support auto-detection of of python interpreters on windows
When specifying the Python interpreter to use for a Python-based
package, the `LIBFOO_PYTHON_INTERPRETER` configuration key can be used.
While a package can use this to target a specific version (e.g.
`python2.7`), in Windows environments, specific Python interpreter paths
may not be available in an invoked PATH. To provide flexibility for
Windows environments, allow a non-path configuration to automatically
detect and use the installation path of the request interpreter by
searching in the registry hives of the running user.

Signed-off-by: James Knight <james.d.knight@live.com>