completion: drop explicit quirk entries
Removing support for describing available quirks in a Bash completion
script. Quirks are not intended for casual use and should not need to be
maintained inside a completion script. A user can query for available
quirks via the `--help-quirks` option.

This commit aims to reduce maintenance required when quirks are added/
removed, as well as provide flexibility if completion scripts are moved
outside the releng-tool's repository.

Signed-off-by: James Knight <james.d.knight@live.com>