fetch-git: support verify git references
Provide the ability for a developer to configure whether or not a
revision for a Git source should be GPG verified. This commit takes
advantage of Git's `verify-tag` and `verify-commit` calls to help verify
a provided revision. If this option is enabled for a package, a failed
validation will prevent the processing of the package (beyond the
fetch/extraction phase).

Signed-off-by: James Knight <james.d.knight@live.com>