package-manager: properly handle cyclic check across packages
To prevent shadowing the `sorted` function, sorted packages for a cyclic
check were renamed to `sorted_pkgs`; however, a line was missed during
this change. This resulted in cyclic checks failing with a non-helpful
error (i.e. exception on an attempt to loop on a `None` value). This
commit corrects the line change.

[1]: 170c63370b9031f6576adba7a26e709b050b0d89

Signed-off-by: James Knight <james.d.knight@live.com>