support implicit ignore-cache fetching
Adjusting the processing of "ignore cache" configurations to be
implicit. This provides users the ability to explicit request either
ignoring cache or not, while also providing flexibility for the
releng-tool engine to decide in selected cases if ignoring a cache may
be desired.

This commit only removes the default-explicit "do not ignore cache"
state. Existing implementation will still default to this state, but
provides flexibility in the future for certain fetch options to handle
their fetch requests differently.

Signed-off-by: James Knight <james.d.knight@live.com>