fetch-git: refactor git cache validation
Refactor the implementation to use the same validation checks when
validating a package's Git cache as well as Git submodule caches.

This change replaces the use of a full `fsck` operation on a Git cache
for a `rev-parse` check. Ideally, the `rev-parse` call will be simple
enough to perform a basic "is this a Git repository" check and avoiding
the need to invoke an `fsck` call which can take a long time for some
repositories to process.

Signed-off-by: James Knight <james.d.knight@live.com>