io: flag failure for ensure_dir_exists if file already exists
The `ensure_dir_exists` IO call would catch a `os.makedirs`'s call and
check for `errno.EEXIST` to rule out the case where a directory may
already exist at the provided target. However, `errno.EEXIST` may also
be set if the target exists for non-directory types. Correcting the
exists-check to ensure that the existing target is of a directory type.

Signed-off-by: James Knight <james.d.knight@live.com>