io: log execute failure for critical invokes
When a call invokes `execute` with the `critical` flag set, if the call
does not provide any meaningful output on why the command has failed, a
user may not understand why releng-tool has stopped processing (unless
debug mode is enabled). In the event that the critical flag is set,
always output the failed command to the configured standard error
stream before hard-stopping.

Signed-off-by: James Knight <james.d.knight@live.com>