re-enable line-length checks at a desired limit
Line length checks for this project were disabled since it was preferred
to use 80-character limits instead of 79-character limits. It was
unknown at the time that flake8 was configurable for this check.

This commit re-enables the `E501` check with the explicit line-length
configuration of 80-characters. In addition, this commit cleans up a
series of implementations to fit within the enforced limit.

Signed-off-by: James Knight <james.d.knight@live.com>