refactor/correct system path injections
releng-tool will modify the system's `PATH` environment variable to
support loading scripts from either the project's root or various paths
found in the hosts directory. This commit relocates this logic to be
performed in the releng-tool's engine instead of the mainline. This has
the benefit of allowing this feature to be tested.

This commit also corrects the injection of a custom system prefix path
for a project. The option `sysroot_prefix` used to add a custom
container `bin` path was applied before the project's configuration
could configure `sysroot_prefix`. With these changes, the host path is
appended after the project's configuration script is processed.

Signed-off-by: James Knight <james.d.knight@live.com>