tests: fallback on tox directory when finding template base
When testing attempts to find the template base directory, fallback to
using the `TOX_WORK_DIR` environment variable if the template base
cannot be found. This is primarily for Python 2.7.x environments which
may not be able to find the test base if the unit testing is at a stage
where the working directory has been changed.

Signed-off-by: James Knight <james.d.knight@live.com>