tests: disable short unit test descriptions
Avoid using short unit test descriptions (pulled from docstrings) over
using the raw test names. This is to make it easier to copy-paste a
specific test for filtering when trying to re-run a given test; also,
make the test output consistent between tests which have a docstring
entry and tests which do not.

Signed-off-by: James Knight <james.d.knight@live.com>