tar: only apply force-local when required/supported
The `--force-local` option helps ensure an archive will properly
extract when a colon is used in the archive name; however, not all `tar`
builds support this option. When an external tar command is used,
detect if the force-local option is available before attempting to use.
This ensures that if the first extraction attempt fails (before falling
back on an internal extraction), it fails due to an issue processing the
archive instead of an unsupported argument.

Signed-off-by: James Knight <james.d.knight@live.com>