extract: no longer warn on empty hash file on extraction
When detecting an empty hash file for an internal package, a warning is
generated to inform a user/developer that no hash checks are being
performed on the downloaded asset. The same warning was generated also
during the extraction phase where the contents of extracted files could
be hashed checked (e.g. for license checks). It is not really a warning
case if a package does not define hash checks for extracted contents,
therefore, adjusting the warning to a verbose message instead.

Signed-off-by: James Knight <james.d.knight@live.com>