engine-scripts: ensure target dir exists before install script invoke
When running releng-tool, the project's "target" directory may never be
created (since not all projects are required to even use a target
directory). However, if a package defines an "install" script for
itself, it is most likely that a target directory is desired. If an
install script is about to be invoked, ensure the "target" directory
exists ahead of time. This can help improve the flexibility of install
scripts which may want to rely on the existence of the target directory,
instead of managing that requirements themselves.

Signed-off-by: James Knight <james.d.knight@live.com>