extract: use internal tar for host-targets without local-only support
Not all tar commands support the `--force-local` option, which is why
the extraction phase was changed [1] to provide the option when it is
available. However, if invoking a tar extraction with a target value
where tar thinks it is an external host and the force-local option
cannot be set, the extract phase may cause the tar command to extract
against an external host. Instead, if a tar target for an external host
is detected, always use the internal tar extraction implementation.

[1]: 7e459758d0232c31e862fb4f305f82de17c27314

Signed-off-by: James Knight <james.d.knight@live.com>