introduce releng_cat script helper
Adding a new script helper call `releng_cat` which a project can use to
easily output file contents to the configured standard output stream.
This can be used to help sanity check the contents of a file when
working in intermediate working environments which may not longer exist
after the releng-tool process completes.

Signed-off-by: James Knight <james.d.knight@live.com>