introduce releng_ls script helper
Adding a new script helper call `releng_ls` which a project can use to
easily output a directory's content to the configured standard output
stream. This can be used to help sanity check the contents of a
directory when working in intermediate working environments which may
no longer exist after the releng-tool process completes.

Signed-off-by: James Knight <james.d.knight@live.com>