Documentation: describe extract tool override formats
While a user can specify extraction tool overrides, the documentation
was not clear on how a tool can accept the file to extract and the
directory to extract into. Adding to the example output that the
`{file}` and `{dir}` options can be used when formatting the final
extraction command that is invoked.

Signed-off-by: James Knight <james.d.knight@live.com>