refactor package key constants
Adjusting the package key constants from specific variables to
attributes of an `Rpk` class. By storing them in an enumerated class
type, a sanity check can be added to verify that each defined key is
registered into the package manager. In addition, this helps minimize
the import list of each individual key.

Signed-off-by: James Knight <james.d.knight@live.com>