tests: improve ability to find the test base
The `find_test_base` call would try to fallback onto `TOX_WORK_DIR`
when the use of `__file__` does not provide an absolute path when using
a Python 2.7.x interpreter. If invoking tox in an environment where the
working directory is changed outside the working directory, a test
environment will not be able to find the test base.

Instead, directly pass in the `{toxinidir}` tox configuration variable
into a `TOX_INI_DIR` environment variable -- and use this to help find
the test base for a fallback case.

Signed-off-by: James Knight <james.d.knight@live.com>