util-io: provide a forced dir destination argument for copy utility
The copy helper (e.g. `releng_copy`) provides a couple of ways for a
developer to copy files/directories in different ways -- specifically,
for files, it provides a way for a caller to copy a file to a specific
file target or a directory target (depending on the arguments and
destination type). In general, a developer can copy file-to-file in a
way such as follows:

    releng_copy('file1', 'file2')

And if desired, copy a file into a directory:

    releng_copy('file1', 'subdir/')
     (or)
    releng_mkdir('subdir')
    releng_copy('file1', 'subdir')

To avoid the need for a caller to explicitly trail the destination path
with a trailing slash or explicitly invoke `releng_mkdir` to ensure the
directory is created, a new argument has been added to provide a hint
that the destination should be a directory. For example:

    releng_copy('file1', 'subdir', dst_dir=True)

Signed-off-by: James Knight <james.d.knight@live.com>