main: allow users to easily inject environment variables
Provides a way for users to easily inject environment variables using
the command line interface. While users can use either export/set in
their environments, or (on Unix) prefix a command call with variable
assignments; these ways to do this is not consistent. A desire was to
allow users to be able to inject environment options in a way similar to
how Makefile calls can perform variable assignments through the command
line. For example:

    make TEST=1

This commit enables such a capability by scanning the argument options
of a releng-tool invoke and applying key-value pair entries into the
working environment. Before, users would have to perform something such
as the following:

    (*nix)
    MY_CUSTOM_ENV=testing releng-tool
    (*nix; alternative)
    export MY_CUSTOM_ENV=testing
    releng-tool
    (windows)
    set MY_CUSTOM_ENV=testing
    releng-tool

With this change, the following will work for any environment:

    releng-tool MY_CUSTOM_ENV=testing

Signed-off-by: James Knight <james.d.knight@live.com>