package-manager: allow injected variables to override package variables
In select build environments, a developer may wish to override specific
package variables on-demand (e.g. override a site, overriding a
revision, etc. -- without having to desire to manage specific override
scripts). With the recent introduction to accept key-value pairs in the
command line, releng-tool will consider these options when processing
package variables. For example, if a package script defines a revision
such as follows:

    TEST_PACKAGE_REVISION = 'v1.0'

And a builder wishes to temporarily override this variable for a build,
they can achieve this by invoking the following:

    releng-tool TEST_PACKAGE_REVISION="test-branch"

Signed-off-by: James Knight <james.d.knight@live.com>