package-pipeline: ensure package-specific envs do not leak
If a package specifies a configuration which should override some
environment variables, the environment is updated before invoking the
specific stages for a package. For example, if a package specifies a
fixed job count, the environment variable `NJOBS` needs to reflect the
proper count when invoking the package's stages. However, an issue with
the original implementation is that these environment variables would
carry on to the trailing packages/post-processing scripts.

This commit ensures that package-specific environment overrides are
restored back to their original state (either the original value or an
unset environment variable) after completing a specific package's
pipeline stages.

Signed-off-by: James Knight <james.d.knight@live.com>