tests: cleanup tar-exists restore for site-url test
When temporarily forcing the hint that an external tar command is not
available for a test, the cleanup operation to restore the flag will
generate an error in pylint (E0601) since the finally-case could be
invoked before `otes` is set. While this should not happen, moving the
introduction of the `otes` variable outside the try to indicate that the
try-case is not aimed to cover this assignment (this, removing the
pylint error).

Signed-off-by: James Knight <james.d.knight@live.com>