fetch-git: correct quick-fetch check
The "quick-fetch" capability relies on using an `ls-remote` call to
sanity check if a reference exists before attempting to fetch. While the
`ls-remote` call can help verify the existence of a desired reference,
the original implementation did not properly invoke this Git method.

First, `ls-remote` calls will return a good error code even if a
provided reference does not exist -- an indication that the call
succeeded in communicating to the server. The `--exit-code` argument
has been added to ensure a non-zero return code is returned when a
provided reference cannot be found. This prevents the ensured "fatal"
logs generated on initial Git fetch attempts.

Second, is has been observed that using `ls-remote` to search for, for
example, a tag, will not work unless adding the explicit reference
parent path (`refs/tags`); adjusting.

Signed-off-by: James Knight <james.d.knight@live.com>