engine: always perform fetch stage for internal pkgs in local-srcs mode
Always ignore a package's fetch-flag state for internal packages when in
a "local sources" mode. The fetch stage for packages in this state help
ensure the local package exists. If this is skipped and a local package
is moved (or originally run in regular mode), releng-tool attempts to
operate in a non-existent directory and fails (with a non-descriptive
error). To ensure a consistent user experience, we will always check
that the directory exists before continuing.

Signed-off-by: James Knight <james.d.knight@live.com>