packages: introduce build "base" directory environment variable
When a package defines a sub-directory to work with, the working
directory for packages and output environment variables will be
populated with values using the configured sub-directory. This can
require developers trying to manually work with relative paths of the
configured sub-directory if they wish to reference content of a package
at the root or between the root and the configured sub-directory. This
commit add a new environment variable `PKG_BUILD_BASE_DIR` to provide
developers an easy way to reference the root a package's extracted
content.

Signed-off-by: James Knight <james.d.knight@live.com>