adjust automatic job count to logical cores
A releng-tool request will attempt to automatically use additional cores
if possible on a system. This logic relies on a
`multiprocessing.cpu_count()` call for a target; however, this value may
be a representation of the total number of logical cores on a system.

Adjusting this calculation to (try to) determine the physical core count
using `os.sched_getaffinity`. This call is not available in all systems/
interpreters, so a halved valued `cpu_count` will be used as an
alternative.

Note that the `psutil` package may be used in the future for a more
consistent physical core count, but attempting to avoid additional
libraries at this time.

Signed-off-by: James Knight <james.d.knight@live.com>