cmake: always provide a parallel job count
Always provide a job count value for the `parallel` argument for a CMake
build. By not specifying a job count, CMake will default to a
tool-specified limit. In some build tools, such as make, this may be an
unlimited count which may not be ideal for more build environments.
Defaulting to the recommended job count for a run should reduce the risk
of overloading a build system.

Signed-off-by: James Knight <james.d.knight@live.com>