introduce releng_copy_into script helper
Provide a new script utility function `releng_copy_into`. This is to
provide an alternative and more function-named-explicit way a developer
can use to copy content into a directory. Avoids a need, when utilizing
`releng_copy`, to ensure either the destination directory exists,
ensuring the destination includes a trailing path separator or avoids
the need to utilize the `dst_dir` argument.

Signed-off-by: James Knight <james.d.knight@live.com>