set clean-related enviro flags when pkg-specific clean is performed
A releng-tool's running environment will set a `RELENG_CLEAN` variable
when a clean-related action is performed. However, if a package-specific
clean operation is made, the flag will not be set. This commit adjusts
the environment initialization to ensure both `RELENG_CLEAN` and
`RELENG_DISTCLEAN` flags are set when a package-specific equivelent is
also configured.

Signed-off-by: James Knight <james.d.knight@live.com>