correct libfoo-exec action tracking
The addition of `libfoo-exec` [1], the `target_action_exec` attribute
is used to track the command to execute. Its assignment uses
`subaction_val` from the for loop used to find the action type, instead
of using the `self.pkg_action` value. In most cases, the implementation
appears to work since the last package action iterated to typically does
not match `PkgAction.EXEC`; however, the order on which package action
types are iterated is not guaranteed. This causes a random chance for
an exec action to be interpreted from a provided non-subaction package.
This commit corrects this by checking for the exec type on the assigned
`self.pkg_action` attribute.

[1]: 0dd2674f46ed96ed809ee67d0aa520832b601f77

Signed-off-by: James Knight <james.d.knight@live.com>