improve file:// package handling on windows
For packages using the `file://` URI sites, releng-tool can fail on
Windows if a user does not properly escape the provided paths. To be
more flexible on how a user prepares a file-based site, always convert
these site values to POSIX paths (which requires no escaping and is also
supported on Windows).

Signed-off-by: James Knight <james.d.knight@live.com>