suppress root warning when running in a container
Running as root is not advised for building, since we would like to help
avoid a user from destroying their host environment. When running in a
container environment, it is more likely that the container is built
with a single user (UID: 0), and this warning will always be generated
when operating in these cases. To help avoid the need for builders to
setup bypasses in these cases, try to detect if we are in a container.
If so, avoid printing this message.

Signed-off-by: James Knight <james.d.knight@live.com>