support specific development modes
Upgrade the development mode capability to allow users to specify a
mode of specific source revisions to fetch. Originally, when enabling
development mode, packages specifying a `LIBFOO_DEVMODE_REVISION` value
which would override which revision to fetch for a site, but this
feature was somewhat limited -- this would only support one other set
of revisions that could be fetched, and there was not a graceful way to
change the site URL (if needed) without some conditional logic in a
package's definition.

This commit improves development mode by allowing a caller to configure
multiple revisions and site entries for a package. Users can then
specify a specific mode they would like to operate in, and this value
can be used to hint a specific revisions/sites to use for a package.
This can be helpful for packages with multiple revision targets (e.g.
branches) that may be in active development.

Signed-off-by: James Knight <james.d.knight@live.com>