engine: suppress corrupt local-srcs conf is applying a new conf
When loading a local sources configuration file, the engine would report
if there was an issue parsing the file. This would occur whether or not
the running active was being configured with a new local sources
configuration. Tweaking the implementation to not report this error
when applying a new change (since this action should put this file back
in a good state again).

Signed-off-by: James Knight <james.d.knight@live.com>