python: allow compiled python host packages by default
Python packages were forced to have the `--no-compile` option set for
all packages. This is less than ideal for host-based packages which will
not be able to take advantage of pre-compiled scripts.

This tweak also allows a package to override a Python package's
argument, permitting `--no-compile` to be manually injected if
applicable to a project's use case.

Signed-off-by: James Knight <james.d.knight@live.com>