perform operation mode management before project script loading
When configuring operation modes, such as development mode or local
sources modem, are intended to be quickly set and the process is
stopped to confirm the configuration of a mode for following runs. The
original mode processing would occur after loading a project's
configuration, which may include project-specific tasks (e.g cleaning up
build resources) that may not be intended to be performed during mode
changes. Instead of preparing flags to indicate a mode-change event,
perform all mode processing before a configuration is parsed.

Signed-off-by: James Knight <james.d.knight@live.com>