allow configuration script fallback for make/scons
Provides the ability for developers to prepare a configuration script to
use for make and SCons packages, instead of utilizing package
bootloaders or supporting a custom configuration target. This can be
useful if a, for example, Make-based project uses the make command for
all stages but configuration, where instead a custom configuration
script to invoke is available.

Signed-off-by: James Knight <james.d.knight@live.com>