removing deprecated releng namespace
The releng namespace has been deprecated for some time (at least seven
releases), so its removal now is as good a time as any.

With its removal, it will allow a next release to officially promote
alternative extension variables for configuration files (since there
will no longer be a conflict with the root configuration file and the
removed `releng` module).

Signed-off-by: James Knight <james.d.knight@live.com>