tool-git: ignore backported configparser when parsing a configuration
The existence of `read_string` may pass in a Python 2.7 environment if
an environment has installed a backported ConfigParser implementation.
This can lead to issue processing in Python 2.7 with a string-based
configuration file that should be a unicode type. Adjusting the check
to explicitly check for Python v3 when using `read_string`, otherwise
to fallback on a string-io instance.

Signed-off-by: James Knight <james.d.knight@live.com>