engine: report success state of cleanup actions
When a cleanup/force build is triggered, various directories or files
will be flagged for removal. If an error occurs when trying to remove a
resource, an error message will print but the error state will not be
reflected in the exit code of the process. Tweaking the implementation
to return a non-zero error code in these cases.

Signed-off-by: James Knight <james.d.knight@live.com>