util-io: provide a forced dir destination argument for move utility
The move helper (e.g. `releng_move`) provides a couple of ways for a
developer to move files/directories in different ways -- specifically,
for files, it provides a way for a caller to move a file to a specific
file target or a directory target (depending on the arguments and
destination type). In general, a developer can move file-to-file in a
way such as follows:

    releng_move('file1', 'file2')

And if desired, move a file into a directory:

    releng_move('file1', 'subdir/')
     (or)
    releng_mkdir('subdir')
    releng_move('file1', 'subdir')

To avoid the need for a caller to explicitly trail the destination path
with a trailing slash or explicitly invoke `releng_mkdir` to ensure the
directory is created, a new argument has been added to provide a hint
that the destination should be a directory. For example:

    releng_move('file1', 'subdir', dst_dir=True)

Signed-off-by: James Knight <james.d.knight@live.com>