introduce releng_move_into script helper
Provide a new script utility function `releng_move_into`. This is to
provide an alternative and more function-named-explicit way a developer
can use to move content into a directory. Avoids a need, when utilizing
`releng_move`, to ensure either the destination directory exists,
ensuring the destination includes a trailing path separator or avoids
the need to utilize the `dst_dir` argument.

Signed-off-by: James Knight <james.d.knight@live.com>