tests: provide a relaxed matching invoke
Adding an option for unit testing to return a zero-exit value when a
provided test pattern cannot find any unit tests. This is primarily to
help run coverage testing on a subset of tests, without having coverage
fail if a unit test suit does not have the same set of unit tests
available as another suite.

Signed-off-by: James Knight <james.d.knight@live.com>