setup.cfg: ensure manifest files are processed for path-based installed
When using a pip to install this package from a Git repository/path, not
all manifest files may be included for older pip installations (for
older interpreters). To be flexible in these environments, hint to
setuptools to ensure package data is included (which should ensure all
MANIFEST.in files are included).

Signed-off-by: James Knight <james.d.knight@live.com>