tweak setuptools build_meta test to following
When attempting to verify setuptools's PEP 517 support, the project
definition used in the test does not match the example provided in
setuptool's documentation [1]. It has been observed in some builds that
this example project may not be processed correctly, since it relies
fully on project definitions inside the `pyproject.toml` file. Since
the test attempts to focus on just validating the use of setuptool's
`setuptools.build_meta` backend, change the example to following the
documentation. This should provide consistent results when using
supported version of setuptools, and re-focus the test to validating
that the backend can invoke as expected.

[1]: https://setuptools.pypa.io/en/latest/build_meta.html

Signed-off-by: James Knight <james.d.knight@live.com>