require a 'local' directory for local vcs-type packages
When a package definition specified a `local` VCS type, sources for the
package could be found directly inside the definition directory. This
can be confusing and have possible collisions mixing up source content
with package definition content. Instead, this change promotes the use
of creating a folder called `local` inside the definition directory,
which in turn will hold all local sources for a package.

For the interim, if the `local` folder is not found, releng-tool will
fallback at looking into the package's definition folder to allow
projects upgrade over time.

Signed-off-by: James Knight <james.d.knight@live.com>