cmake: correct missing find locations for target installs
A package which defaults to a `target` install did not have various
include/library paths and prefixes for the target area registered for
CMake projects. This was a result of a "default" install type (`None`)
not being treated as a `target` install type when a CMake configuration
is processed. This commit corrects this be properly checking the
populated installation type when making decisions on which paths and
prefixes to register.

Signed-off-by: James Knight <james.d.knight@live.com>