cmake: remove unused variable suppression
The original idea of adding variable suppression for CMake definitions
was to mask various common configuration options that releng-tool sets,
but may not be applicable to a given package. This results in having a
notice for each CMake-configuration of the suppression attempt, as well
as does not provide a way for developers to easily interpret which
CMake-options are being considered. This commit removes the suppression
to give developers the standard warning information, allowing them to
get the best feedback for a CMake configuration event.

Signed-off-by: James Knight <james.d.knight@live.com>