registry: avoid reloading modules
The call to `load_module` will force reload a given module into the
interpreter. This can lead to unexpected results with usages of `super`.
There is no need to reload any module requested by an extension load.
Instead, if a module is detected to be already loaded, use the loaded
reference instead.

Signed-off-by: James Knight <james.d.knight@live.com>