engine-patch: only notify when patch stage is relevant
Originally, the stage processing for a package would also notify that
the patch stage was being processed, even if there were not explicit
patches to apply. In most cases, projects may never take advantage of
patches, so such a notify is not always helpful. Instead, only note that
patching is performed if there is a patch script to invoke or patches to
apply. This not only cleans up non-patching projects, but also provides
an explicit hint to users when a project is performing patching.

Signed-off-by: James Knight <james.d.knight@live.com>