engine-patch: provide package-defined directory support for patches
This commit improves a developers ability to tailor packages for either
the root of the extract sources (default; new) or into a sub-directory
which can be configured via `LIBFOO_PATCH_SUBDIR`. This prevents issues
for users attempting to patch a project when using `LIBFOO_BUILD_SUBDIR`
but wish to patch from the root of the extracted content.

Signed-off-by: James Knight <james.d.knight@live.com>