registry: generate extension exception when not providing a class type
While registry safe-guarding helped sanity check for a class's callable
methods needed for various extension types, it did not explicitly check
if the provided handle was itself a class. This could lead to less
graceful exception states if a non-class type was provided. This commit
adds an explicit check for the class type, and generates an exception
when a non-class is detected.

Signed-off-by: James Knight <james.d.knight@live.com>