promote alternative explicit-module execution style
This tool provides a module to allow execution when using either
`releng-tool` or `releng_tool` when providing Python a module to load.
The purpose of `releng-tool` was to provide a module format that matched
the `releng-tool` helper executable (and somewhat also the pip
installation project). Since support for this has been available for the
last two versions, switching to promote the dash-styled module.

Signed-off-by: James Knight <james.d.knight@live.com>