correct assignments of "reinvoked-stage" flags
When a user forces a redo of a stage (e.g. `pkg-reconfigure`), various
flags are configured a package's script environment to indicate that
a force redo of a stage is occurring. For example, when reconfiguring,
`RELENG_RECONFIGURE` is set. It should be expected that when performing
a full reconfiguration of a package (and not just a reconfig-only
request) that other trailing flags are set -- i.e. `RELENG_REBUILD` and
`RELENG_REINSTALL`. This is however, not the case.

This commit tweaks the setting of these flags to ensure trailing stage
flags are properly set for an action (e.g. a rebuild request will also
set a reinstall request flag).

Signed-off-by: James Knight <james.d.knight@live.com>