stats: provide support do disable pdf generation by quirk
The PDF statistics generation will only create PDFs if matplotlib is
installed. There may be cases where even if the running interpreter
supports this module, a user may wish to not generate such a file.

This commit introduced a `releng.stats.no_pdf` quirk to help prevent any
chance of generating a PDF.

Signed-off-by: James Knight <james.d.knight@live.com>